Parsing CSV into codetest Database ( MySQL) which is then outputting Countries and amount of people in then via JSON hash.
This involved first of all defining a schema and creating two models for the tables, Places and Person.
I then connected to the MySQL  database with activerecord, clearing all existing data in the tables. I then used the CSV library to parse the CSV files into the tables.
I linked the tables via a Foreign Key, this enables me to access the place of a person.
Finally i output the data to JSON by selecting all the People, then finding the assocaited place of theese people. I then group the countires of these places and count
the values within them.